+++ packages/plugin/src/manager/load/pkg.test.ts
@@ -1,6 +1,6 @@
 import t from 'tap'
 
-import { resolvePackageEntry } from './pkg.js'
+import { resolveBrowserEntry, resolvePackageEntry } from './pkg.js'
 
 t.test('resolvePackageEntry', async t => {
   t.test('main', async t => {
@@ -23,5 +23,27 @@ t.test('resolvePackageEntry', async t => {
 
     t.same(entry, 'dist/index.js')
   })
+})
 
+t.test('resolveBrowserEntry', async t => {
+  t.test('browser', async t => {
+    const entry = resolveBrowserEntry({browser: 'index.js'})
+
+
+    t.ok(entry)
+  })
+
+  t.test('only main should return falsy', async t => {
+    const entry = resolveBrowserEntry({main: 'index.js'})
+
+
+    t.notOk(entry)
+  })
+
+  t.test('exports.node.module', async t => {
+    const entry = resolveBrowserEntry({exports: {node: {module: 'dist/index.mjs'}, 'browser': 'dist/public/index.js'}})
+
+
+    t.same(entry, 'dist/public/index.js')
+  })
 })