+++ packages/plugin/src/manager/types.ts
@@ -1,7 +1,14 @@
 import { TPlugin } from '@home-gallery/types'
 
 export type TPluginContext = {
-  file: string
   plugin: TPlugin
+  /**
+   * Plugin file. Null if plugin was added by source
+   */
+  file: string | null
+  /**
+   * Public plugin Dir. Null if plugin is a single file
+   */
+  publicDir: string | null
   initialized: boolean
 }